+++ src/extension.ts
@@ -1,12 +1,11 @@
 import * as vscode from 'vscode';
-
-import checkRequiredExtentions from './check_required_extensions';
 import { updateRustAnalyzerConfig } from './update_cargo_args';
-import { cargoRunner } from './cargo-runner';
-import { taskProvider } from './task-provider';
+import { cargo_runner_config, cargoRunner } from './cargo-runner';
 import { getOutputChannel, log } from './logger';
-import workspaceConfig from './workspace_config';
+import { taskProvider } from './tasks';
+import { MissingExtension } from './errors';
 
+let rustAnalyzerLoading = true;
 
 export async function activate(context: vscode.ExtensionContext) {
 	await checkRequiredExtentions();
@@ -16,9 +15,45 @@ export async function activate(context: vscode.ExtensionContext) {
 	context.subscriptions.push(cargoRunner);
 	context.subscriptions.push(updateRustAnalyzerConfig);
 	context.subscriptions.push(vscode.workspace.onDidChangeConfiguration(() => {
-		log(JSON.stringify(workspaceConfig()), 'debug');
+		log(JSON.stringify(cargo_runner_config()), 'debug');
 	}));
 
+}
+
+function loadExtensions(): Promise<boolean> {
+	return new Promise((resolve, reject) => {
+		const rustAnalyzer = vscode.extensions.getExtension('rust-lang.rust-analyzer');
+		const codeLLDB = vscode.extensions.getExtension('vadimcn.vscode-lldb');
+
+		if (!rustAnalyzer?.isActive || !codeLLDB?.isActive) {
+			let extensions: string[] = [];
+			if (!rustAnalyzer) {
+				extensions.push('rust-lang.rust-analyzer');
+			}
+			if (!codeLLDB) {
+				extensions.push('vadimcn.vscode-lldb');
+			}
+			const message = `Missing extension${extensions.length > 1 ? 's' : ''}:\n ${extensions.join(' and ')}`;
+			log(message, 'error');
+			throw new MissingExtension(message);
+		} else {
+			resolve(false);
+		}
+	});
+}
+
+async function checkRequiredExtentions() {
+	try {
+		setTimeout(async () => {
+			rustAnalyzerLoading = await loadExtensions();
+		}, 1000);
+	}
+	catch (error: unknown) {
+		if (error instanceof MissingExtension) {
+			vscode.window.showErrorMessage(error.message);
+			return;
+		}
+	}
 }
 
 export function deactivate() { }
