+++ src/handle_custom_bench.ts
@@ -1,9 +1,8 @@
 import * as path from 'path';
-import { findCargoToml } from './find_cargo_toml';
-import getCargoToml from './get_cargo_toml';
 import { log } from './logger'; 
 import { getDocument } from './editor';
 import * as vscode from 'vscode';
+import { findCargoToml, parseCargoManifest } from './cargo_manifest';
 
 export default function handleCustomBench(runner: vscode.CodeLens) {
 	if (
@@ -16,7 +15,7 @@ export default function handleCustomBench(runner: vscode.CodeLens) {
 			return;
 		}
 
-		const cargo = getCargoToml(cargoTomlPath ?? '');
+		const cargo = parseCargoManifest(cargoTomlPath ?? '');
 		if (!cargo?.bench?.length) {
 			log('No benches found in Cargo.toml', 'debug');
 			return;
