--- src/task-provider.ts
@@ -1,7 +1,0 @@
-import * as vscode from 'vscode';
-import { CargoRunnerTaskProvider } from './tasks';
-
-export 	const taskProvider = vscode.tasks.registerTaskProvider(
-    CargoRunnerTaskProvider.cargoType,
-    new CargoRunnerTaskProvider()
-);
